Split it to packages for better maintenance and resusability
This commit also adds support to invoke a command which is useful
in cases where there's no function available in kubernetes client
to use.
