+++ setup.cfg
@@ -15,7 +15,7 @@ zip_safe = False
 packages = find:
 include_package_data = True
 package_dir =
-    =src
+    =cerberus
 # Add here dependencies of your project (semicolon/line-separated), e.g.
 install_requires = PyYAML
 # tests_require = pytest; pytest-cov
